property_schema.py renamed to property_data.py, in an attempt to align filenames with class names. Next up: try to decommission location_params() since distinguishing get_property() from get_properties() should handle what it used to handle.
